🚚 Move GitHubWiki interfaces into their own file
